requirements.txt
streamlit
Pillow
ultralytics
opencv-python
pandas
numpy
sentence-transformers
requests
beautifulsoup4
lxml
